fix: ensure searchPokemon works in both development and production

+++ src/services/pokemon-service.ts
@@ -1,42 +1,74 @@
 import Pokemon from "../models/pokemon";
+import POKEMONS from "../models/mock-pokemon";
 
 export default class PokemonService {
-  static pokemons: Pokemon[] = [];
+
+  static pokemons: Pokemon[] = POKEMONS;
+
+  static isDev = (!process.env.NODE_ENV || process.env.NODE_ENV === 'development');
 
   static getPokemons(): Promise<Pokemon[]> {
-    return fetch('http://localhost:3001/pokemons')
-      .then(response => response.json())
-      .then(data => {
-        this.pokemons = data; // Stocke les Pokémon dans la propriété de classe
-        return data;
-      })
-      .catch(error => this.handleError(error));
+    if(this.isDev) {
+      return fetch('http://localhost:3001/pokemons')
+        .then(response => response.json())
+        .then(data => {
+          this.pokemons = data; // Stocke les Pokémon dans la propriété de classe
+          return data;
+        })
+        .catch(error => this.handleError(error));
+    }
+
+    return new Promise(resolve => {
+      resolve(this.pokemons);
+    });
   }
 
   static getPokemon(id: number): Promise<Pokemon | null> {
-    return fetch(`http://localhost:3001/pokemons/${id}`)
-      .then(response => response.json())
-      .then(data => this.isEmpty(data) ? null : data)
-      .catch(error => this.handleError(error));
+    if(this.isDev) {
+      return fetch(`http://localhost:3001/pokemons/${id}`)
+        .then(response => response.json())
+        .then(data => this.isEmpty(data) ? null : data)
+        .catch(error => this.handleError(error));
+    }
+
+    return new Promise(resolve => {
+      const pokemon = this.pokemons.find(pokemon => id === pokemon.id);
+      resolve(pokemon ? pokemon : null);
+    });
   }
 
   static updatePokemon(pokemon: Pokemon): Promise<Pokemon> {
-    return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
-      method: 'PUT',
-      body: JSON.stringify(pokemon),
-      headers: { 'Content-Type': 'application/json' }
-    })
-      .then(response => response.json())
-      .catch(error => this.handleError(error));
+    if(this.isDev) {
+      return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
+        method: 'PUT',
+        body: JSON.stringify(pokemon),
+        headers: { 'Content-Type': 'application/json' }
+      })
+        .then(response => response.json())
+        .catch(error => this.handleError(error));
+    }
+
+    return new Promise(resolve => {
+      const index = this.pokemons.findIndex(p => p.id === pokemon.id);
+      this.pokemons[index] = pokemon;
+      resolve(pokemon);
+    });
   }
 
   static deletePokemon(pokemon: Pokemon): Promise<{}> {
-    return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
-      method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' }
-    })
-      .then(response => response.json())
-      .catch(error => this.handleError(error));
+    if(this.isDev) {
+      return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
+        method: 'DELETE',
+        headers: { 'Content-Type': 'application/json' }
+      })
+        .then(response => response.json())
+        .catch(error => this.handleError(error));
+    }
+
+    return new Promise(resolve => {
+      this.pokemons = this.pokemons.filter(p => p.id !== pokemon.id);
+      resolve({});
+    });
   }
 
   static addPokemon(pokemon: Pokemon): Promise<Pokemon> {
@@ -44,17 +76,30 @@ export default class PokemonService {
       delete pokemon.created;
     }
 
-    return fetch(`http://localhost:3001/pokemons/`, {
-      method: 'POST',
-      body: JSON.stringify(pokemon),
-      headers: { 'Content-Type': 'application/json' }
-    })
-      .then(response => response.json())
-      .catch(error => this.handleError(error));
+    if(this.isDev) {
+      return fetch(`http://localhost:3001/pokemons`, {
+        method: 'POST',
+        body: JSON.stringify(pokemon),
+        headers: { 'Content-Type': 'application/json' }
+      })
+        .then(response => response.json())
+        .catch(error => this.handleError(error));
+    }
+
+    return new Promise(resolve => {
+      this.pokemons.push(pokemon);
+      resolve(pokemon);
+    });
   }
 
   static searchPokemon(term: string): Promise<Pokemon[]> {
-    return new Promise((resolve) => {
+    if(this.isDev) {
+      return fetch(`http://localhost:3001/pokemons?q=${term}`)
+        .then(response => response.json())
+        .catch(error => this.handleError(error));
+    }
+
+    return new Promise(resolve => {
       const filteredPokemons = this.pokemons.filter(pokemon => 
         pokemon.name.toUpperCase().includes(term.toUpperCase())
       );
